refactor & fix: router로 분리, 댓글 등록 엔드포인트 수정 등
