+++ .env
@@ -1,2 +1,2 @@
 DATABASE_URL="postgresql://hyungmincho:112211@localhost:5432/panda?schema=public"
-PORT=3000
+PORT=5200
