+++ prisma/schema.prisma
@@ -13,14 +13,6 @@ datasource db {
   url      = env("DATABASE_URL")
 }
 
-// model User {
-//   id        String   @id @default(uuid())
-//   nickname  String   @unique
-//   createdAt DateTime @default(now())
-//   updatedAt DateTime @updatedAt
-//   Article      Article[]
-// }
-
 model Product {
   id          String    @id @default(uuid())
   name        String
@@ -53,3 +45,11 @@ model Comment {
   product   Product? @relation(fields: [productId], references: [id], onDelete: Cascade)
   productId String?
 }
+
+// model User {
+//   id        String   @id @default(uuid())
+//   nickname  String   @unique
+//   createdAt DateTime @default(now())
+//   updatedAt DateTime @updatedAt
+//   Article      Article[]
+// }
